Move git repository inside the Assets Folder
Making a submodule of the project doesn't work very well because the user
would end up copying all of the top-level unity project files.
Additionally, all of these files are not needed anyway and it is simpler
to move the repo to where the Raconteur files are.

Also added a .gitattributes file for the Twine examples folder.